fix(ci): resolve 30 mypy errors, add preflight validation gates
- Fix no-any-return across 19 Phase 3/4 physics modules (np.asarray wraps)
- Fix return-value, callable type, var-annotated errors
- Add EpedPedestalModel class for integrated_transport_solver import
- numpy compat: np.trapezoid fallback to np.trapz
- Add python-compat and hash-pin gates to preflight.py
- Add 'abl' (ablation) to typos dictionary
- Increase RL training test timeout to 300s
- Fix GEMINI_PHASE6_REPORT.md missing trailing newline

+++ src/scpn_control/core/elm_model.py
@@ -62,7 +62,7 @@ class PeelingBallooningBoundary:
         j_norm = max(0.0, j_edge / j_crit)
         a_norm = max(0.0, alpha_edge / a_crit)
 
-        return 1.0 - np.sqrt(j_norm**2 + a_norm**2)
+        return float(1.0 - np.sqrt(j_norm**2 + a_norm**2))
 
 
 @dataclass
